Split addr module
